requirements.txt
mesa~=3.1.1
networkx~=3.4.2
streamlit~=1.41.1
matplotlib~=3.10.0
pillow~=11.0.0